kidney_label_classifier/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2019.11.28=py36_0
cffi=1.14.0=py36h2e261b9_0
cuda90=1.0=h6433d27_0
cycler=0.10.0=py36h93f1223_0
dbus=1.13.12=h746ee38_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.63.1=h5a9c865_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
intel-openmp=2020.0=166
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=py36he6710b0_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
matplotlib=3.1.3=py36_0
matplotlib-base=3.1.3=py36hef1b27d_0
mkl=2020.0=166
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
ncurses=6.2=he6710b0_0
ninja=1.9.0=py36hfd86e86_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
olefile=0.46=py_0
openssl=1.1.1d=h7b6447c_4
pandas=1.0.1=py36h0573a6f_0
pcre=8.43=he6710b0_0
pillow=6.1.0=py36h34e0f95_0
pip=20.0.2=py36_1
pycparser=2.19=py_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.10=h0371630_0
python-dateutil=2.8.1=py_0
pytorch=1.0.0=py3.6_cuda9.0.176_cudnn7.4.1_1
pytz=2019.3=py_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
scikit-learn=0.22.1=py36hd81dba3_0
scipy=1.4.1=py36h0b6359f_0
setuptools=45.2.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.14.0=py36_0
sqlite=3.31.1=h7b6447c_0
tk=8.6.8=hbc83047_0
torchvision=0.2.1=py_2
tornado=6.0.3=py36h7b6447c_3
wheel=0.34.2=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
